requirements.txt
Flask>=2.3.0
python-dotenv>=1.0.0
py-ecc>=6.0.0
ecdsa>=0.18.0

